EDITOR-#420 auto change add/edit button
